# swagger-docs/swagger.json
{
    "swagger": "2.0",
    "paths": {
        "/orderschedule/index.php/Home/Accounts/signin": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "\u767b\u5f55",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "\u7528\u6237\u767b\u5f55\u8d26\u53f7",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "\u767b\u5f55\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Accounts/logout": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "\u767b\u51fa",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Accounts/signup": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "\u6ce8\u518c",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "userName",
                        "in": "query",
                        "description": "\u7528\u6237\u767b\u5f55\u8d26\u53f7",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "\u767b\u5f55\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "\u7528\u6237\u59d3\u540d",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "\u7528\u6237\u624b\u673a\u53f7",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "\u7528\u6237\u5730\u5740",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Accounts/changePassword": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "\u4fee\u6539\u5bc6\u7801",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "originPassword",
                        "in": "query",
                        "description": "\u65e7\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "newPassword",
                        "in": "query",
                        "description": "\u65b0\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "rePassword",
                        "in": "query",
                        "description": "\u786e\u8ba4\u65b0\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Businesses/establish": {
            "post": {
                "tags": [
                    "Businesses"
                ],
                "summary": "\u65b0\u5efa\u4e1a\u52a1\u7c7b\u578b",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "\u4e1a\u52a1\u540d\u79f0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "\u4e1a\u52a1\u7b80\u8ff0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "\u4e1a\u52a1\u7c7b\u578b",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/orderschedule/index.php/Home/Businesses/obtainAll": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "\u83b7\u53d6\u6240\u6709\u4e1a\u52a1\u7c7b\u578b",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
...
    },
...
}